moved audio playback to separate component
